+++ utility.h
@@ -1,7 +1,8 @@
 #ifndef RMS_UTILITY_H
 #define RMS_UTILITY_H
+#include <stdint.h>
 
 void flush_buffer(void);
-int generate_uuid(int length);
+uint64_t generate_uuid(int length);
 
 #endif //RMS_UTILITY_H
